modularize styles and rollback bootstrap to get glyphicon support
